backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.13
psycopg2-binary==2.9.6
asyncpg==0.28.0
alembic==1.11.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Validation and Serialization
pydantic==2.9.0
pydantic-settings==2.6.1
pydantic-extra-types==2.10.0
email-validator==2.2.0

# HTTP and API
httpx==0.24.1
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Monitoring
structlog==23.1.0

# CORS and Middleware
starlette==0.27.0

# Environment and Configuration
python-dotenv==1.0.0

# Optional: Redis for caching
redis==4.6.0
aioredis==2.0.1

# Optional: Email validation
email-validator==2.2.0
